Expose UnitJsonSerializer and UnitJsonDeserializer
There's no reason to hide those classes.
